requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
gTTS>=2.5.0
deep-translator>=1.11.0